backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7

Pillow==10.0.1
django-filter==23.3
markdown==3.5.1
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0 